+++ contracts/mocks/ERC721Mock.sol
@@ -1,30 +1,18 @@
 pragma solidity ^0.4.24;
 
 import "../token/ERC721/ERC721.sol";
-import "../token/ERC721/ERC721Mintable.sol";
-import "../token/ERC721/ERC721Burnable.sol";
 
 
 /**
  * @title ERC721Mock
- * This mock just provides a public mint and burn functions for testing purposes,
- * and a public setter for metadata URI
+ * This mock just provides a public mint and burn functions for testing purposes
  */
-contract ERC721Mock is ERC721, ERC721Mintable, ERC721Burnable {
-  constructor(string name, string symbol) public
-    ERC721Mintable()
-    ERC721(name, symbol)
-  {}
-
-  function exists(uint256 tokenId) public view returns (bool) {
-    return _exists(tokenId);
+contract ERC721Mock is ERC721 {
+  function mint(address to, uint256 tokenId) public {
+    _mint(to, tokenId);
   }
 
-  function setTokenURI(uint256 tokenId, string uri) public {
-    _setTokenURI(tokenId, uri);
-  }
-
-  function removeTokenFrom(address from, uint256 tokenId) public {
-    _removeTokenFrom(from, tokenId);
+  function burn(uint256 tokenId) public {
+    _burn(ownerOf(tokenId), tokenId);
   }
 }
