+++ contracts/token/ERC721/ERC721Full.sol
@@ -1,0 +1,19 @@
+pragma solidity ^0.4.24;
+
+import "./ERC721.sol";
+import "./ERC721Enumerable.sol";
+import "./ERC721Metadata.sol";
+
+
+/**
+ * @title Full ERC721 Token
+ * This implementation includes all the required and some optional functionality of the ERC721 standard
+ * Moreover, it includes approve all functionality using operator terminology
+ * @dev see https://github.com/ethereum/EIPs/blob/master/EIPS/eip-721.md
+ */
+contract ERC721Full is ERC721, ERC721Enumerable, ERC721Metadata {
+  constructor(string name, string symbol) ERC721Metadata(name, symbol)
+    public
+  {
+  }
+}
