+++ contracts/token/ERC721/IERC721Metadata.sol
@@ -1,0 +1,14 @@
+pragma solidity ^0.4.24;
+
+import "./IERC721.sol";
+
+
+/**
+ * @title ERC-721 Non-Fungible Token Standard, optional metadata extension
+ * @dev See https://github.com/ethereum/EIPs/blob/master/EIPS/eip-721.md
+ */
+contract IERC721Metadata is IERC721 {
+  function name() external view returns (string name);
+  function symbol() external view returns (string symbol);
+  function tokenURI(uint256 tokenId) public view returns (string);
+}
