+++ test/token/ERC721/ERC721Burnable.test.js
@@ -1,4 +1,4 @@
-const { shouldBehaveLikeERC721Basic } = require('./ERC721Basic.behavior');
+const { shouldBehaveLikeERC721 } = require('./ERC721.behavior');
 const {
   shouldBehaveLikeMintAndBurnERC721,
 } = require('./ERC721MintBurn.behavior');
@@ -17,6 +17,6 @@ contract('ERC721Burnable', function ([_, creator, ...accounts]) {
     this.token = await ERC721Burnable.new({ from: creator });
   });
 
-  shouldBehaveLikeERC721Basic(creator, minter, accounts);
+  shouldBehaveLikeERC721(creator, minter, accounts);
   shouldBehaveLikeMintAndBurnERC721(creator, minter, accounts);
 });
